Gallery2: Migrate from switch to if-else for menu items
`case` requires final variables which here R.* are not.

+++ src/com/android/gallery3d/filtershow/presets/UserPresetsAdapter.java
@@ -136,14 +136,11 @@ public class UserPresetsAdapter extends ArrayAdapter<Action>
 
     @Override
     public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.editView:
-                v.requestFocus();
-                break;
-            case R.id.deleteUserPreset:
-                Action action = (Action) v.getTag();
-                deletePreset(action);
-                break;
+        if (v.getId() == R.id.editView) {
+            v.requestFocus();
+        } else if (v.getId() == R.id.deleteUserPreset) {
+            Action action = (Action) v.getTag();
+            deletePreset(action);
         }
     }
